// FRONT/src/app/interfaces/origination/project-area/project-area.interface.ts
export interface ProjectArea {
  Id_project_area:          number;
  project_id:               number;
  id_certification:         number;
  total_area:               string;
  achurado:                 string;
  expropriated_area:        string;
  file_path_phina:          string;
  total_perimeter_area_ran: string;
  ran_application:          number;
  internal_flat_surface:    string;
  expropriated_areas:       string;
  plan_yaer:                number;
  file_path_internal_flat:  string;
  id_ap_por_sig:            number;
  file_path_ap:             string;
  observations_ap:          string;
  Date:                     string;
  NameUser:                 string;
}